Extract LineReplacementEvaluator to its own file
